// src/app/shared/interface/cliente.interface.ts
export interface Cliente {
    fullName?: string;
    email?: string;
    telefone?: string;
    endereco?: string;
    cidade?: string;
    estado?: string;
    password?: string;
    phrase?: string; 
    CorpName?: string;
    cargo?: string;
    nivel?: string;
    admissao?: string;
    ferramenta?: string;
}